Refactor - move RegistrationSubmissionsController_part.cs into main file
